requirements.txt
supabase
python-dotenv
pandas
streamlit